Struct spectrum1_explorer::PawsNotifySpectrumUseRequest

pub struct PawsNotifySpectrumUseRequest {
    pub device_desc: Option<DeviceDescriptor>,
    pub type_: Option<String>,
    pub location: Option<GeoLocation>,
    pub version: Option<String>,
    pub spectra: Vec<SpectrumMessage>,
}

The spectrum-use notification message which must contain the geolocation of the Device and parameters required by the regulatory domain.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

notify spectrum use paws (request)

Fields

device_desc

Device descriptor information is required in the spectrum-use notification message.

type_

The message type (e.g., INIT_REQ, AVAIL_SPECTRUM_REQ, ...).

Required field.

location

The geolocation of the master device (the device that is sending the spectrum-use notification) to the database is required in the spectrum-use notification message.

version

The PAWS version. Must be exactly 1.0.

Required field.

spectra

A spectrum list is required in the spectrum-use notification. The list specifies the spectrum that the device expects to use, which includes frequency ranges and maximum power levels. The list may be empty if the device decides not to use any of spectrum. For consistency, the psdBandwidthHz value should match that from one of the spectrum elements in the corresponding available spectrum response previously sent to the device by the database. Note that maximum power levels in the spectrum element must be expressed as power spectral density over the specified psdBandwidthHz value. The actual bandwidth to be used (as computed from the start and stop frequencies) may be different from the psdBandwidthHz value. As an example, when regulatory rules express maximum power spectral density in terms of maximum power over any 100 kHz band, then the psdBandwidthHz value should be set to 100 kHz, even though the actual bandwidth used can be 20 kHz.