Struct spectrum1_explorer::PawsNotifySpectrumUseResponse

pub struct PawsNotifySpectrumUseResponse {
    pub kind: Option<String>,
    pub version: Option<String>,
    pub type_: Option<String>,
}

An empty response to the notification.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

notify spectrum use paws (response)

Fields

kind

Identifies what kind of resource this is. Value: the fixed string "spectrum#pawsNotifySpectrumUseResponse".

version

The PAWS version. Must be exactly 1.0.

Required field.

type_

The message type (e.g., INIT_REQ, AVAIL_SPECTRUM_REQ, ...).

Required field.